moved the proxy details to the proxyinfo modal
